Add support for using the HTTPSecureSocket class
This pushes back the minimum Xojo version needed to use the project (without RB-libcURL or plugins) to 2014r3.

Starting in that version we can use the HTTPSecureSocket to talk to OpenAI. This is the version that updated the supported SSL/TLS versions (and OpenAI requires a newish version).

Not all operations are supported, only GET and POST. Presently, only OpenAI.File.Delete() and OpenAi.FineTune.Cancel() are affected by this limitation.