Fix IllegalCastException when receiving an API permission error
Project scoped API keys can have restricted permissions, which when encountered will raise an error. However the format of the error object is different from other API errors, which results in an IllegalCastException when we try to access the error info. This commit changes the OpenAIException to account for this, allowing the error details to be read.